refactor: address review feedback — naming, tests, comments
- Rename camunda.log.event_time_min/max → camunda.event.time_min/max
- Add schema URL to otelMeter for consistency with otelLogger
- Update performance comment with JMH numbers (~0.2ms for 50 metrics)
- Merge OtelSdkManagerMetricTest into OtelSdkManagerTest via @Nested
- Merge overlapping handler counter tests into one
- Rename shouldRoundTripAllSequenceNumbers → shouldSerializeAndDeserializeAllFields
- Move helper methods before @Nested classes (checkstyle InnerTypeLast)
- Add metricSequenceNumber assertion to AnalyticsExporterTest
