+++ zeebe/exporters/analytics-exporter/src/main/java/io/camunda/exporter/analytics/OtelSdkManager.java
@@ -75,7 +75,8 @@ public class OtelSdkManager {
             .build();
     otelLogger =
         sdk.getLogsBridge().loggerBuilder(INSTRUMENTATION_SCOPE).setSchemaUrl(SCHEMA_URL).build();
-    otelMeter = sdk.getMeterProvider().get(INSTRUMENTATION_SCOPE);
+    otelMeter =
+        sdk.getMeterProvider().meterBuilder(INSTRUMENTATION_SCOPE).setSchemaUrl(SCHEMA_URL).build();
     registerExportWindowGauge();
     return this;
   }
@@ -106,8 +107,8 @@ public class OtelSdkManager {
     metricWindow.record(position, eventTimeMs);
   }
 
-  // Performance: collectAndExport() runs on the partition thread. Benchmarked at ~8ms for 50
-  // metrics × 1999 dimensions, ~236ms for 1000 × 1999 (worst case). If collection time becomes
+  // Performance: collectAndExport() runs on the partition thread. Benchmarked at ~0.2ms for 50
+  // metrics × 1999 dimensions, ~5ms for 1000 × 1999 (worst case). If collection time becomes
   // a concern, offload to a background thread: create a fresh MeterProvider per flush ("seal &
   // create"), swap atomically, and collect the old provider on a background ExecutorService.
   void flushMetrics() {
