+++ zeebe/exporters/analytics-exporter/src/test/java/io/camunda/exporter/analytics/OtelSdkManagerTest.java
@@ -7,18 +7,33 @@
  */
 package io.camunda.exporter.analytics;
 
+import static io.camunda.exporter.analytics.AnalyticsAttributes.EVENT_TIME_MAX;
+import static io.camunda.exporter.analytics.AnalyticsAttributes.EVENT_TIME_MIN;
+import static io.camunda.exporter.analytics.AnalyticsAttributes.LOG_POSITION_END;
+import static io.camunda.exporter.analytics.AnalyticsAttributes.LOG_POSITION_START;
+import static io.camunda.exporter.analytics.AnalyticsAttributes.METRIC_EXPORT_WINDOW;
+import static io.camunda.exporter.analytics.AnalyticsAttributes.METRIC_SEQUENCE_NUMBER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import io.opentelemetry.api.common.AttributeKey;
+import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.logs.data.LogRecordData;
 import io.opentelemetry.sdk.logs.export.LogRecordExporter;
+import io.opentelemetry.sdk.metrics.data.LongPointData;
+import io.opentelemetry.sdk.metrics.data.MetricData;
+import io.opentelemetry.sdk.testing.exporter.InMemoryLogRecordExporter;
+import io.opentelemetry.sdk.testing.exporter.InMemoryMetricReader;
 import java.util.Collection;
+import java.util.Optional;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 
@@ -305,5 +320,195 @@ class OtelSdkManagerTest {
         return CompletableResultCode.ofSuccess();
       }
     };
+  }
+
+  @Nested
+  class MetricRecording {
+
+    private InMemoryMetricReader metricReader;
+    private OtelSdkManager manager;
+
+    @BeforeEach
+    void setUp() {
+      metricReader = InMemoryMetricReader.create();
+      manager =
+          TestOtelSdkManager.inMemoryWithMetrics(InMemoryLogRecordExporter.create(), metricReader);
+    }
+
+    @Test
+    void shouldRecordMetricByName() {
+      // when
+      manager.incrementMetric("test.counter", 100L, 1000L, Attributes.empty());
+
+      // then
+      assertThat(findMetric(metricReader.collectAllMetrics(), "test.counter")).isPresent();
+    }
+
+    @Test
+    void shouldReuseCounterForSameName() {
+      // when
+      manager.incrementMetric("test.counter", 100L, 1000L, Attributes.empty());
+      manager.incrementMetric("test.counter", 100L, 1000L, Attributes.empty());
+
+      // then
+      assertThat(findMetric(metricReader.collectAllMetrics(), "test.counter"))
+          .hasValueSatisfying(
+              metric -> {
+                final long total =
+                    metric.getLongSumData().getPoints().stream()
+                        .mapToLong(LongPointData::getValue)
+                        .sum();
+                assertThat(total).isEqualTo(2);
+              });
+    }
+
+    @Test
+    void shouldCreateSeparateCountersForDifferentNames() {
+      // when
+      manager.incrementMetric("counter.a", 100L, 1000L, Attributes.empty());
+      manager.incrementMetric("counter.b", 200L, 2000L, Attributes.empty());
+
+      // then
+      final var metrics = metricReader.collectAllMetrics();
+      assertThat(findMetric(metrics, "counter.a")).isPresent();
+      assertThat(findMetric(metrics, "counter.b")).isPresent();
+    }
+
+    @Test
+    void shouldPreserveDimensionAttributes() {
+      // given
+      final var attrs =
+          Attributes.of(
+              AttributeKey.stringKey("process.id"), "order", AttributeKey.longKey("version"), 3L);
+
+      // when
+      manager.incrementMetric("test.counter", 100L, 1000L, attrs);
+
+      // then
+      assertThat(findMetric(metricReader.collectAllMetrics(), "test.counter"))
+          .hasValueSatisfying(
+              metric ->
+                  assertThat(metric.getLongSumData().getPoints())
+                      .first()
+                      .satisfies(
+                          point -> {
+                            assertThat(
+                                    point.getAttributes().get(AttributeKey.stringKey("process.id")))
+                                .isEqualTo("order");
+                            assertThat(point.getAttributes().get(AttributeKey.longKey("version")))
+                                .isEqualTo(3L);
+                          }));
+    }
+
+    @Test
+    void shouldEmitExportWindowGaugeWithMetadata() {
+      // given
+      manager.incrementMetric("test.counter", 100L, 5000L, Attributes.empty());
+      manager.incrementMetric("test.counter", 200L, 6000L, Attributes.empty());
+
+      // when
+      final var metrics = metricReader.collectAllMetrics();
+
+      // then
+      assertThat(findMetric(metrics, METRIC_EXPORT_WINDOW))
+          .isPresent()
+          .hasValueSatisfying(
+              metric ->
+                  assertThat(metric.getLongGaugeData().getPoints())
+                      .first()
+                      .satisfies(
+                          point -> {
+                            assertThat(point.getValue()).isEqualTo(2);
+                            final var pointAttrs = point.getAttributes();
+                            assertThat(pointAttrs.get(METRIC_SEQUENCE_NUMBER)).isEqualTo(1L);
+                            assertThat(pointAttrs.get(LOG_POSITION_START)).isEqualTo(100L);
+                            assertThat(pointAttrs.get(LOG_POSITION_END)).isEqualTo(200L);
+                            assertThat(pointAttrs.get(EVENT_TIME_MIN)).isEqualTo(5000L);
+                            assertThat(pointAttrs.get(EVENT_TIME_MAX)).isEqualTo(6000L);
+                          }));
+    }
+
+    @Test
+    void shouldIncrementFlushSequenceAcrossCollections() {
+      // given — two collection cycles
+      manager.incrementMetric("test.counter", 100L, 1000L, Attributes.empty());
+      metricReader.collectAllMetrics(); // flush_sequence = 1
+
+      manager.incrementMetric("test.counter", 200L, 2000L, Attributes.empty());
+
+      // when
+      final var metrics = metricReader.collectAllMetrics();
+
+      // then
+      assertThat(findMetric(metrics, METRIC_EXPORT_WINDOW))
+          .isPresent()
+          .hasValueSatisfying(
+              metric ->
+                  assertThat(metric.getLongGaugeData().getPoints())
+                      .first()
+                      .satisfies(
+                          point ->
+                              assertThat(point.getAttributes().get(METRIC_SEQUENCE_NUMBER))
+                                  .isEqualTo(2L)));
+    }
+
+    @Test
+    void shouldResetWindowAfterCollection() {
+      // given
+      manager.incrementMetric("test.counter", 100L, 5000L, Attributes.empty());
+      metricReader.collectAllMetrics(); // resets window
+
+      manager.incrementMetric("test.counter", 300L, 8000L, Attributes.empty());
+
+      // when
+      final var metrics = metricReader.collectAllMetrics();
+
+      // then — should reflect only the second event
+      assertThat(findMetric(metrics, METRIC_EXPORT_WINDOW))
+          .isPresent()
+          .hasValueSatisfying(
+              metric ->
+                  assertThat(metric.getLongGaugeData().getPoints())
+                      .first()
+                      .satisfies(
+                          point -> {
+                            assertThat(point.getValue()).isEqualTo(1);
+                            final var pointAttrs = point.getAttributes();
+                            assertThat(pointAttrs.get(LOG_POSITION_START)).isEqualTo(300L);
+                            assertThat(pointAttrs.get(LOG_POSITION_END)).isEqualTo(300L);
+                            assertThat(pointAttrs.get(EVENT_TIME_MIN)).isEqualTo(8000L);
+                            assertThat(pointAttrs.get(EVENT_TIME_MAX)).isEqualTo(8000L);
+                          }));
+    }
+
+    @Test
+    void shouldEmitHeartbeatGaugeWhenNoEvents() {
+      // when — no incrementMetric calls
+      final var metrics = metricReader.collectAllMetrics();
+
+      // then — gauge still emitted as heartbeat with event_count=0
+      assertThat(findMetric(metrics, METRIC_EXPORT_WINDOW))
+          .isPresent()
+          .hasValueSatisfying(
+              metric ->
+                  assertThat(metric.getLongGaugeData().getPoints())
+                      .first()
+                      .satisfies(
+                          point -> {
+                            assertThat(point.getValue()).isZero();
+                            final var pointAttrs = point.getAttributes();
+                            assertThat(pointAttrs.get(METRIC_SEQUENCE_NUMBER)).isEqualTo(1L);
+                            // Sentinels must not leak — only sequence_number present
+                            assertThat(pointAttrs.get(LOG_POSITION_START)).isNull();
+                            assertThat(pointAttrs.get(LOG_POSITION_END)).isNull();
+                            assertThat(pointAttrs.get(EVENT_TIME_MIN)).isNull();
+                            assertThat(pointAttrs.get(EVENT_TIME_MAX)).isNull();
+                          }));
+    }
+
+    private Optional<MetricData> findMetric(
+        final Collection<MetricData> metrics, final String name) {
+      return metrics.stream().filter(m -> m.getName().equals(name)).findFirst();
+    }
   }
 }
